improve tracing import

+++ scripts/import-tracing-runtime.sh
@@ -30,7 +30,6 @@ PATHS_TO_GIT=(
   primitives\\/account\\/
   primitives\\/rpc\\/txpool
   primitives\\/xcm\\/
-  relay-encoder
 )
 
 declare -A SHARED_PATHS
@@ -42,19 +41,27 @@ SHARED_PATHS["..\/evm_tracer"]="..\/..\/..\/shared\/runtime\/evm_tracer"
 SPEC_VERSION=$1
 GIT_REF=${2:-"runtime-$SPEC_VERSION"}
 
-# Get moonbeam repository shopshot
-echo "Get moonbeam branch..."
-rm -rf tmp
-mkdir tmp
-git clone https://github.com/PureStake/moonbeam -b $GIT_REF tmp/moonbeam
+if [[ "$SPEC_VERSION" == "local" ]]; then
+  MOONBEAM_PATH="../../"
+  GIT_REV="$GIT_REF"
+else
+  MOONBEAM_PATH="tmp/moonbeam"
+  GIT_REV="runtime-$SPEC_VERSION"
+
+  # Get moonbeam repository snapshot
+  echo "Get moonbeam snapshot..."
+  rm -rf tmp
+  mkdir tmp
+  git clone https://github.com/PureStake/moonbeam --depth 1 -b $GIT_REF $MOONBEAM_PATH
+fi
 
 # Copy relevant files
 echo "Copy relevant files and folders..."
 mkdir -p tracing/$SPEC_VERSION/runtime
-cp tmp/moonbeam/Cargo.lock tracing/$SPEC_VERSION/Cargo.lock
-cp tmp/moonbeam/rust-toolchain tracing/$SPEC_VERSION/rust-toolchain
-cp -r tmp/moonbeam/runtime/common tracing/$SPEC_VERSION/runtime/
-cp -r tmp/moonbeam/runtime/moon* tracing/$SPEC_VERSION/runtime/
+cp $MOONBEAM_PATH/Cargo.lock tracing/$SPEC_VERSION/Cargo.lock
+cp $MOONBEAM_PATH/rust-toolchain tracing/$SPEC_VERSION/rust-toolchain
+cp -r $MOONBEAM_PATH/runtime/common tracing/$SPEC_VERSION/runtime/
+cp -r $MOONBEAM_PATH/runtime/moon* tracing/$SPEC_VERSION/runtime/
 
 # Remove irrelevant files
 rm -rf tracing/$SPEC_VERSION/runtime/relay-encoder
@@ -75,10 +82,11 @@ done
 # Replace some path dependencies by git dependencies
 echo "Replace path dependencies by git dependencies..."
 for PATH_TO_GIT in ${PATHS_TO_GIT[@]}; do
-  sed -i -e "s/path = \"..\/..\/$PATH_TO_GIT\"/git = \"https:\/\/github.com\/purestake\/moonbeam\", rev = \"runtime-$SPEC_VERSION\"/g" tracing/$SPEC_VERSION/runtime/common/Cargo.toml
+  sed -i -e "s/path = \"..\/..\/$PATH_TO_GIT\"/git = \"https:\/\/github.com\/purestake\/moonbeam\", rev = \"$GIT_REV\"/g" tracing/$SPEC_VERSION/runtime/common/Cargo.toml
   for CHAIN in ${CHAINS[@]}; do
-    sed -i -e "s/path = \"..\/..\/$PATH_TO_GIT\"/git = \"https:\/\/github.com\/purestake\/moonbeam\", rev = \"runtime-$SPEC_VERSION\"/g" tracing/$SPEC_VERSION/runtime/$CHAIN/Cargo.toml
-    sed -i -e "s/path = \"..\/$PATH_TO_GIT\"/git = \"https:\/\/github.com\/purestake\/moonbeam\", rev = \"runtime-$SPEC_VERSION\"/g" tracing/$SPEC_VERSION/runtime/$CHAIN/Cargo.toml
+    sed -i -e "s/path = \"..\/..\/$PATH_TO_GIT\"/git = \"https:\/\/github.com\/purestake\/moonbeam\", rev = \"$GIT_REV\"/g" tracing/$SPEC_VERSION/runtime/$CHAIN/Cargo.toml
+    sed -i -e "s/path = \"..\/relay-encoder\"/git = \"https:\/\/github.com\/purestake\/moonbeam\", rev = \"$GIT_REV\"/g" tracing/$SPEC_VERSION/runtime/$CHAIN/Cargo.toml
+    sed -i -e 's/moonbeam-rpc-primitives-debug = { path = "..\/..\/primitives\/rpc\/debug", default-features = false }/moonbeam-rpc-primitives-debug = { path = "..\/..\/primitives\/rpc\/debug", default-features = false, features = \[ "transaction_v0" \] }/g' runtime/moonbase/Cargo.toml tracing/$SPEC_VERSION/runtime/$CHAIN/Cargo.toml
   done
 done
